+++ src/commands/Admin/conf.js
@@ -8,7 +8,7 @@ module.exports = class extends Command {
 			permLevel: 6,
 			guarded: true,
 			description: (msg) => msg.language.get('COMMAND_CONF_SERVER_DESCRIPTION'),
-			usage: '<set|get|reset|list|remove> [key:string] [value:string] [...]',
+			usage: '<get|set|remove|reset|list> [key:string] [value:string] [...]',
 			usageDelim: ' '
 		});
 	}
@@ -18,6 +18,11 @@ module.exports = class extends Command {
 		if (['set', 'remove'].includes(action) && value.length === 0) throw msg.language.get('COMMAND_CONF_NOVALUE');
 
 		return this[action](msg, key, value);
+	}
+
+	get(msg, key) {
+		const { path } = this.client.gateways.guilds.getPath(key, { avoidUnconfigurable: true, piece: true });
+		return msg.sendMessage(msg.language.get('COMMAND_CONF_GET', path.path, path.resolveString(msg)));
 	}
 
 	async set(msg, key, valueToSet) {
@@ -33,11 +38,6 @@ module.exports = class extends Command {
 	async reset(msg, key) {
 		const { path } = await msg.guild.configs.reset(key, true);
 		return msg.sendMessage(msg.language.get('COMMAND_CONF_RESET', path.path, path.resolveString(msg)));
-	}
-
-	get(msg, key) {
-		const { path } = this.client.gateways.guilds.getPath(key, { avoidUnconfigurable: true, piece: true });
-		return msg.sendMessage(msg.language.get('COMMAND_CONF_GET', path.path, path.resolveString(msg)));
 	}
 
 	list(msg, key) {