+++ src/events/configUpdateEntry.js
@@ -3,11 +3,11 @@ const { Event } = require('klasa');
 module.exports = class extends Event {
 
 	run(configs) {
-		if (!this.client.sharded) return;
+		if (!this.client.shard) return;
 		if (configs.type === 'users') {
 			this.client.shard.broadcastEval(`
 				if (this.shard.id !== ${this.client.shard.id}) {
-					const user = this.users.get(${configs.id});
+					const user = this.users.get('${configs.id}');
 					if (user) user.configs.sync();
 				}
 				`);