+++ src/lib/structures/MonitorStore.js
@@ -60,7 +60,7 @@ class MonitorStore extends Collection {
 	 * Deletes a monitor from the store
 	 * @since 0.0.1
 	 * @param {Monitor|string} name The monitor object or a string representing the structure this store caches
-	 * @return {boolean} whether or not the delete was successful.
+	 * @returns {boolean} whether or not the delete was successful.
 	 */
 	delete(name) {
 		const piece = this.resolve(name);