+++ src/lib/structures/ProviderStore.js
@@ -59,18 +59,18 @@ class ProviderStore extends Collection {
 	/**
 	 * The default provider
 	 * @since 0.5.0
-	 * @return {Provider} The default provider set in KlasaClientOptions.provider
+	 * @returns {Provider} The default provider set in KlasaClientOptions.providers
 	 * @readonly
 	 */
 	get default() {
-		return this.get(this.client.options.provider.engine);
+		return this.get(this.client.options.providers.default);
 	}
 
 	/**
 	 * Deletes a provider from the store
 	 * @since 0.0.1
 	 * @param {Provider|string} name The provider object or a string representing the structure this store caches
-	 * @return {boolean} whether or not the delete was successful.
+	 * @returns {boolean} whether or not the delete was successful.
 	 */
 	delete(name) {
 		const pro = this.resolve(name);