+++ src/lib/util/constants.js
@@ -23,13 +23,18 @@ exports.DEFAULTS = {
 		language: 'en-US',
 		prefix: '!',
 		preserveConfigs: true,
-		provider: { engine: 'json' },
 		readyMessage: (client) => `Successfully initialized. Ready to serve ${client.guilds.size} guilds.`,
 		typing: false,
 		customPromptDefaults: {
 			promptTime: 30000,
 			promptLimit: Infinity,
 			quotedStringSupport: false
+		},
+		providers: { default: 'json' },
+		gateways: {
+			guilds: {},
+			users: {},
+			clientStorage: {}
 		},
 		pieceDefaults: {
 			commands: {